refactor(Profit): 수익률 계산 로직 -> Profit 객체 안에 구현
